feat : update Course Map #26
여행 코스 내용을 수정할 수 있습니다.
